backend/pyproject.toml
[project]
name = "daily-pilot"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai==1.86.0",
    "langgraph==0.4.8",
    "langchain==0.3.26",
    "streamlit==1.46.0",
    "python-dotenv==1.1.0",
    "fastapi[standard]==0.115.13",
    "pydantic-settings==2.9.1",
    "SQLAlchemy==2.0.41",
    "asyncpg>=0.29.0",
    "redis==6.2.0",
    "motor==3.7.1",
    "greenlet==3.2.3",
    "psycopg2==2.9.10",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.ruff]
line-length = 120
target-version = "py310"
include = [
    "tests/**/*.py",
]
extend-include = [".venv", "build", "dist", "migrations"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]  # E, F: flake8 기본, I: isort, UP: pyupgrade

[tool.ruff.format]
line-ending = "lf"
quote-style = "double"
skip-magic-trailing-comma = false
docstring-code-format = true

